Moved markdown tips to its own component. Cleaned up. Switched to a textarea for the bio, but user can't edit it, nor is it mirrored anywhere.
