+++ liwords-ui/src/profile/profile.tsx
@@ -1,5 +1,4 @@
 import React, { useCallback, useEffect } from 'react';
-import ReactMarkdown from 'react-markdown';
 import { Link, useParams, useLocation } from 'react-router-dom';
 import { notification, Card, Table, Row, Col, Switch } from 'antd';
 import axios, { AxiosError } from 'axios';
